requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Mail>=0.9.1
Werkzeug>=2.3.7
SQLAlchemy>=2.0.23
email-validator>=2.1.0
python-dotenv>=1.0.0
reportlab>=4.0.7
xlsxwriter>=3.1.2
pandas>=2.1.3
Pillow>=10.1.0
qrcode>=7.4.2
APScheduler>=3.10.4
Flask-Bcrypt>=1.0.1
Flask-Admin>=1.6.1
Flask-RESTful>=0.3.9
Flask-Cors>=3.0.10
Flask-JWT-Extended>=4.4.4
Flask-SocketIO>=5.3.2
Flask-Assets>=2.0
Flask-Babel>=2.0.0
Flask-Caching>=1.10.1
Flask-Compress>=1.13
Flask-DebugToolbar>=0.13.1
Flask-Limiter>=3.3.0
Flask-Marshmallow>=0.14.0
Flask-Principal>=0.4.0
Flask-Security-Too>=5.1.2
Flask-Session>=0.4.0
gunicorn>=20.1.0
python-dateutil>=2.8.2
openpyxl>=3.1.2
bcrypt>=3.2.0
